@@ -1,29 +1,28 @@
 /*
-    Copyright 2013-2017 appPlant GmbH
+ Copyright 2013 Sebastián Katzer
 
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
+ Licensed to the Apache Software Foundation (ASF) under one
+ or more contributor license agreements.  See the NOTICE file
+ distributed with this work for additional information
+ regarding copyright ownership.  The ASF licenses this file
+ to you under the Apache License, Version 2.0 (the
+ "License"); you may not use this file except in compliance
+ with the License.  You may obtain a copy of the License at
 
-     http://www.apache.org/licenses/LICENSE-2.0
+ http://www.apache.org/licenses/LICENSE-2.0
 
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
+ Unless required by applicable law or agreed to in writing,
+ software distributed under the License is distributed on an
+ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ KIND, either express or implied.  See the License for the
+ specific language governing permissions and limitations
+ under the License.
  */
 
 package de.appplant.cordova.plugin.background;
 
 import android.app.Activity;
 import android.content.ComponentName;
-import android.content.Context;
 import android.content.Intent;
 import android.content.ServiceConnection;
 import android.os.IBinder;
@@ -31,19 +30,20 @@ import android.os.IBinder;
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaPlugin;
 import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
+
+import de.appplant.cordova.plugin.background.ForegroundService.ForegroundBinder;
+
+import static android.content.Context.BIND_AUTO_CREATE;
+import static de.appplant.cordova.plugin.background.BackgroundModeExt.clearKeyguardFlags;
 
 public class BackgroundMode extends CordovaPlugin {
 
     // Event types for callbacks
-    private enum Event {
-        ACTIVATE, DEACTIVATE, FAILURE
-    }
+    private enum Event { ACTIVATE, DEACTIVATE, FAILURE }
 
     // Plugin namespace
-    private static final String JS_NAMESPACE =
-            "cordova.plugins.backgroundMode";
+    private static final String JS_NAMESPACE = "cordova.plugins.backgroundMode";
 
     // Flag indicates if the app is in background or foreground
     private boolean inBackground = false;
@@ -61,17 +61,18 @@ public class BackgroundMode extends CordovaPlugin {
     private ForegroundService service;
 
     // Used to (un)bind the service to with the activity
-    private final ServiceConnection connection = new ServiceConnection() {
+    private final ServiceConnection connection = new ServiceConnection()
+    {
         @Override
-        public void onServiceConnected(ComponentName name, IBinder service) {
-            ForegroundService.ForegroundBinder binder =
-                    (ForegroundService.ForegroundBinder) service;
-
+        public void onServiceConnected (ComponentName name, IBinder service)
+        {
+            ForegroundBinder binder = (ForegroundBinder) service;
             BackgroundMode.this.service = binder.getService();
         }
 
         @Override
-        public void onServiceDisconnected(ComponentName name) {
+        public void onServiceDisconnected (ComponentName name)
+        {
             fireEvent(Event.FAILURE, "'service disconnected'");
         }
     };
@@ -85,34 +86,35 @@ public class BackgroundMode extends CordovaPlugin {
      *                 calling back into JavaScript.
      *
      * @return Returning false results in a "MethodNotFound" error.
-     *
-     * @throws JSONException
      */
     @Override
     public boolean execute (String action, JSONArray args,
-                            CallbackContext callback) throws JSONException {
+                            CallbackContext callback)
+    {
+        boolean validAction = true;
 
-        if (action.equalsIgnoreCase("configure")) {
-            JSONObject settings = args.getJSONObject(0);
-            boolean update      = args.getBoolean(1);
-
-            configure(settings, update);
-        }
-        else
-        if (action.equalsIgnoreCase("enable")) {
-            enableMode();
-        }
-        else
-        if (action.equalsIgnoreCase("disable")) {
-            disableMode();
-        }
-        else {
-            BackgroundExt.execute(action, cordova, webView);
+        switch (action)
+        {
+            case "configure":
+                configure(args.optJSONObject(0), args.optBoolean(1));
+                break;
+            case "enable":
+                enableMode();
+                break;
+            case "disable":
+                disableMode();
+                break;
+            default:
+                validAction = false;
         }
 
-        callback.success();
+        if (validAction) {
+            callback.success();
+        } else {
+            callback.error("Invalid action: " + action);
+        }
 
-        return true;
+        return validAction;
     }
 
     /**
@@ -121,10 +123,22 @@ public class BackgroundMode extends CordovaPlugin {
      * @param multitasking Flag indicating if multitasking is turned on for app.
      */
     @Override
-    public void onPause(boolean multitasking) {
-        super.onPause(multitasking);
-        inBackground = true;
-        startService();
+    public void onPause(boolean multitasking)
+    {
+        try {
+            inBackground = true;
+            startService();
+        } finally {
+            clearKeyguardFlags(cordova.getActivity());
+        }
+    }
+
+    /**
+     * Called when the activity is no longer visible to the user.
+     */
+    @Override
+    public void onStop () {
+        clearKeyguardFlags(cordova.getActivity());
     }
 
     /**
@@ -133,8 +147,8 @@ public class BackgroundMode extends CordovaPlugin {
      * @param multitasking Flag indicating if multitasking is turned on for app.
      */
     @Override
-    public void onResume(boolean multitasking) {
-        super.onResume(multitasking);
+    public void onResume (boolean multitasking)
+    {
         inBackground = false;
         stopService();
     }
@@ -143,15 +157,17 @@ public class BackgroundMode extends CordovaPlugin {
      * Called when the activity will be destroyed.
      */
     @Override
-    public void onDestroy() {
-        super.onDestroy();
+    public void onDestroy()
+    {
         stopService();
+        android.os.Process.killProcess(android.os.Process.myPid());
     }
 
     /**
      * Enable the background mode.
      */
-    private void enableMode() {
+    private void enableMode()
+    {
         isDisabled = false;
 
         if (inBackground) {
@@ -162,7 +178,8 @@ public class BackgroundMode extends CordovaPlugin {
     /**
      * Disable the background mode.
      */
-    private void disableMode() {
+    private void disableMode()
+    {
         stopService();
         isDisabled = true;
     }
@@ -173,7 +190,8 @@ public class BackgroundMode extends CordovaPlugin {
      * @param settings The settings
      * @param update A truthy value means to update the running service.
      */
-    private void configure(JSONObject settings, boolean update) {
+    private void configure(JSONObject settings, boolean update)
+    {
         if (update) {
             updateNotification(settings);
         } else {
@@ -186,17 +204,15 @@ public class BackgroundMode extends CordovaPlugin {
      *
      * @param settings The new default settings
      */
-    private void setDefaultSettings(JSONObject settings) {
+    private void setDefaultSettings(JSONObject settings)
+    {
         defaultSettings = settings;
     }
 
     /**
-     * The settings for the new/updated notification.
-     *
-     * @return
-     *      updateSettings if set or default settings
+     * Returns the settings for the new/updated notification.
      */
-    protected static JSONObject getSettings() {
+    static JSONObject getSettings () {
         return defaultSettings;
     }
 
@@ -205,7 +221,8 @@ public class BackgroundMode extends CordovaPlugin {
      *
      * @param settings The config settings
      */
-    private void updateNotification(JSONObject settings) {
+    private void updateNotification(JSONObject settings)
+    {
         if (isBind) {
             service.updateNotification(settings);
         }
@@ -215,21 +232,18 @@ public class BackgroundMode extends CordovaPlugin {
      * Bind the activity to a background service and put them into foreground
      * state.
      */
-    private void startService() {
+    private void startService()
+    {
         Activity context = cordova.getActivity();
 
         if (isDisabled || isBind)
             return;
 
-        Intent intent = new Intent(
-                context, ForegroundService.class);
+        Intent intent = new Intent(context, ForegroundService.class);
 
         try {
-            context.bindService(intent,
-                    connection, Context.BIND_AUTO_CREATE);
-
+            context.bindService(intent, connection, BIND_AUTO_CREATE);
             fireEvent(Event.ACTIVATE, null);
-
             context.startService(intent);
         } catch (Exception e) {
             fireEvent(Event.FAILURE, String.format("'%s'", e.getMessage()));
@@ -242,17 +256,14 @@ public class BackgroundMode extends CordovaPlugin {
      * Bind the activity to a background service and put them into foreground
      * state.
      */
-    private void stopService() {
+    private void stopService()
+    {
         Activity context = cordova.getActivity();
+        Intent intent    = new Intent(context, ForegroundService.class);
 
-        Intent intent = new Intent(
-                context, ForegroundService.class);
-
-        if (!isBind)
-            return;
+        if (!isBind) return;
 
         fireEvent(Event.DEACTIVATE, null);
-
         context.unbindService(connection);
         context.stopService(intent);
 
@@ -265,37 +276,22 @@ public class BackgroundMode extends CordovaPlugin {
      * @param event The name of the event
      * @param params Optional arguments for the event
      */
-    private void fireEvent (Event event, String params) {
-        String eventName;
+    private void fireEvent (Event event, String params)
+    {
+        String eventName = event.name().toLowerCase();
+        Boolean active   = event == Event.ACTIVATE;
 
-        switch (event) {
-            case ACTIVATE:
-                eventName = "activate"; break;
-            case DEACTIVATE:
-                eventName = "deactivate"; break;
-            default:
-                eventName = "failure";
-        }
-
-        String active = event == Event.ACTIVATE ? "true" : "false";
-
-        String flag = String.format("%s._isActive=%s;",
+        String str = String.format("%s._setActive(%b)",
                 JS_NAMESPACE, active);
 
-        String depFn = String.format("%s.on%s(%s);",
-                JS_NAMESPACE, eventName, params);
+        str = String.format("%s;%s.on('%s', %s)",
+                str, JS_NAMESPACE, eventName, params);
 
-        String fn = String.format("%s.fireEvent('%s',%s);",
-                JS_NAMESPACE, eventName, params);
+        str = String.format("%s;%s.fireEvent('%s',%s);",
+                str, JS_NAMESPACE, eventName, params);
 
-        final String js = flag + fn + depFn;
+        final String js = str;
 
-        cordova.getActivity().runOnUiThread(new Runnable() {
-            @Override
-            public void run() {
-                webView.loadUrl("javascript:" + js);
-            }
-        });
+        cordova.getActivity().runOnUiThread(() -> webView.loadUrl("javascript:" + js));
     }
-
 }
