@@ -173,7 +173,6 @@ NSString* const kAPPBackgroundEventDeactivate = @"deactivate";
     // Play music even in background and dont stop playing music
     // even another app starts playing sound
     [session setCategory:AVAudioSessionCategoryPlayback
-             withOptions:AVAudioSessionCategoryOptionMixWithOthers
                    error:NULL];
 
     // Active the audio session
@@ -223,7 +222,7 @@ NSString* const kAPPBackgroundEventDeactivate = @"deactivate";
     NSString* flag = [NSString stringWithFormat:@"%@._isActive=%@;",
                       kAPPBackgroundJsNamespace, active];
 
-    NSString* depFn = [NSString stringWithFormat:@"%@.on%@();",
+    NSString* depFn = [NSString stringWithFormat:@"%@.on('%@');",
                        kAPPBackgroundJsNamespace, event];
 
     NSString* fn = [NSString stringWithFormat:@"%@.fireEvent('%@');",
@@ -242,7 +241,7 @@ NSString* const kAPPBackgroundEventDeactivate = @"deactivate";
  */
 + (NSString*) wkProperty
 {
-    NSString* str = @"X2Fsd2F5c1J1bnNBdEZvcmVncm91bmRQcmlvcml0eQ==";
+    NSString* str = @"YWx3YXlzUnVuc0F0Rm9yZWdyb3VuZFByaW9yaXR5";
     NSData* data  = [[NSData alloc] initWithBase64EncodedString:str options:0];
 
     return [[NSString alloc] initWithData:data encoding:NSUTF8StringEncoding];
@@ -267,7 +266,7 @@ NSString* const kAPPBackgroundEventDeactivate = @"deactivate";
                forKey:[APPBackgroundMode wkProperty]];
 
         [obj setValue:[NSNumber numberWithBool:NO]
-               forKey:@"_requiresUserActionForMediaPlayback"];
+               forKey:@"requiresUserActionForMediaPlayback"];
 
         return obj;
     }
